test: co-locate tests in sibling *_test.py; rename docs/api -> docs/apis
Enforce CLAUDE.md rule 9 (each module foo.py has a sibling foo_test.py) by redistributing the cross-cutting test files to module siblings, and rename the API docs directory.

- docs/api -> docs/apis (+ index.rst / faq.rst / .gitignore refs)
- jansen_rit_test.py -> jansen_rit/{step,tr,network}_test.py
- horn_test.py -> horn/{step,sequential,variants}_test.py
- wilson_cowan_variants_test.py merged into wilson_cowan_test.py
- gradient_test.py split per model (hopf/wilson_cowan/jansen_rit/qif/fhn)
- characterization_test.py goldens inlined into model siblings; drop _characterization_goldens.json
- branch_coverage_test.py split: new _xy_model_test.py, qif/fhn noise, import_compat deprecation

Full suite: 336 passed, 98.17% coverage.
