+++ docs/apis/noise.rst
